requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2
requests>=2.28
pyyaml>=6.0
matplotlib>=3.7
plotly>=5.0
duckdb>=1.0
pyarrow>=14.0
